+++ src/containers/loginpage/loginpage.css
@@ -166,11 +166,11 @@ body {
     }
 
     .login-form button {
-        width: 260px;
+        width: 270px;
     }
 
     .password input {
-        width: 210px;
+        width: 225px;
     }
 
     .password button {
