+++ src/containers/wallet/wallet.css
@@ -26,13 +26,16 @@
     color: #1CB335;
 }
 
-.medisync__wallet_details button {
-    display: block;
+/* .medisync__wallet_details-button button {
+    display: flex;
     margin: 0 auto;
+    width: 450px;
+    height: 50px;
+    text-align: left;
+    justify-content: center;
     font-weight: 500;
     font-size: 16px;
-    line-height: 32px;
-    padding: 10px 80px;
+    padding: 5px 20px;
     color: #FEFAFA;
     background: #436CDF;
     box-shadow: 0px 2px 4px rgba(190, 190, 190, 0.2);
@@ -40,4 +43,4 @@
     border: none;
     outline: none;
     cursor: pointer;
-}
+} */
