+++ src/css/styles.css
@@ -1,7 +1,12 @@
 @media screen and (min-width: 300px) {
-body{
+/* body{
   overflow-x: hidden;
   font-family: 'Poppins', sans-serif;
+} */
+
+body{
+  overflow-x: hidden;
+  font-family: var(--font-family);
 }
 
 /* font-family: 'Poppins', sans-serif; */
@@ -16,7 +21,7 @@ body{
   flex-direction: column;
   gap: 35px;
   width: 20%;
-  padding: 30px 0 0 10px;
+  padding: 30px 0 200px 10px;
   border-right: 2px solid lightgrey;
 }
 
@@ -26,13 +31,14 @@ body{
 }
 
 .logo {
-  width: 90%;
+  width: 105px;
+  padding-left: 15px;
 }
 
 ul {
   display: flex;
   flex-direction: column;
-  gap: 15px;
+  gap: 2px;
   padding: 0;
   margin: 0;
 }
@@ -47,7 +53,7 @@ li {
   align-items: center;
   font-size: 15px;
   font-weight: 600;
-  padding-left: 10px;
+  padding-left: 15px;
   width: 95%;
   text-decoration: none;
   color: black;
@@ -58,7 +64,7 @@ li {
   color: white;
   background-color: #436CDF;
   width: 90%;
-  border-radius: 10px;
+  border-radius: 5px;
 }
 
 section {
@@ -118,14 +124,23 @@ span {
   font-size: 15px;
 }
 
+h3 {
+  font-size: 22px;
+  font-weight: 500;
+  color: #494949;
+  margin-bottom: 0;
+}
+
 select {
   display: flex;
   justify-content: center;
   align-items: center;
   height: 40px;
   width: 180px;
-  font-weight: 600;
-  font-size: 12px;
+  color: #686868;
+  background: #f7f7f7;
+  font-weight: 500;
+  font-size: 14px;
   padding: 5px;
   border-radius: 5px;
 }
@@ -149,6 +164,12 @@ input {
   outline: none;
   border: 1px solid lightgrey;
   padding: 0;
+  color: #686868;
+  background: #f7f7f7;
+  font-weight: 500;
+  font-size: 14px;
+  padding-left: 10px;
+  border-radius: 5px;
 }
 
 .first_section,
@@ -207,6 +228,15 @@ select {
   align-items: center;
   gap: 10px;
   cursor: pointer;
+}
+
+.new-entry_section img {
+  width: 40px;
+}
+
+.new-entry_section h3 {
+  font-size: 18px;
+  margin-top: 0;
 }
 
 .allergy {
